transforming from tinyint to int
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C52310

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SalePriceListDataManager.cs
@@ -145,7 +145,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
                 PriceListId = (int)reader["ID"],
                 OwnerType = (Entities.SalePriceListOwnerType)GetReaderValue<int>(reader, "OwnerType"),
                 EffectiveOn = GetReaderValue<DateTime>(reader, "EffectiveOn"),
-                PriceListType = (SalePriceListType?)GetReaderValue<byte?>(reader, "PriceListType"),
+                PriceListType = (SalePriceListType?)GetReaderValue<int?>(reader, "PriceListType"),
                 ProcessInstanceId = GetReaderValue<long>(reader, "ProcessInstanceID"),
                 FileId = GetReaderValue<long>(reader, "FileID"),
                 CreatedTime = GetReaderValue<DateTime>(reader, "CreatedTime"),
@@ -154,7 +154,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
                 UserId = GetReaderValue<int>(reader, "UserID"),
                 Description = GetReaderValue<string>(reader, "Description"),
                 PricelistStateBackupId = GetReaderValue<long?>(reader, "PricelistStateBackupID"),
-                PricelistSource = (SalePricelistSource?)GetReaderValue<byte?>(reader, "PricelistSource"),
+                PricelistSource = (SalePricelistSource?)GetReaderValue<int?>(reader, "PricelistSource"),
             };
 
             return salePriceList;
